+++ src/modules/role/domain/entity/role-permission-mapping.entity.ts
@@ -1,0 +1,15 @@
+import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
+import { Role } from './role.entity';
+import { Permission } from './permission.entity';
+
+@Entity()
+export class RolePermissionMapping {
+  @PrimaryGeneratedColumn({ type: 'bigint' })
+  identifier: number;
+
+  @ManyToOne(() => Role, { lazy: true, cascade: false })
+  authority_group: Role;
+
+  @ManyToOne(() => Permission, { lazy: true, cascade: false })
+  authority: Permission;
+}
