+++ src/modules/role/domain/entity/role-type.enum.ts
@@ -1,4 +1,4 @@
-export enum AuthorityType {
+export enum RoleType {
   MEMBER = 'member',
   ADMIN = 'admin',
   EDITOR = 'editor',
